backend/requirements.txt
openai==1.55.1
python-dotenv==1.0.1
fastapi
uvicorn
httpx<0.28